Move ``plonelabel`` blueprint to ``yafowil.plone.widgets.label``.
