+++ src/yafowil/plone/plone5.py
@@ -57,7 +57,7 @@ def register_macros():
 ###############################################################################
 
 def register():
-    import common
+    from yafowil.plone.widgets import label
     factory.register_global_preprocessors([plone_preprocessor])
     factory.register_theme('plone5', 'yafowil.plone', resourcedir, css=css)
