+++ src/yafowil/plone/widgets/label.py
